+++ typescript/helloworld/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @hyperlane-xyz/helloworld
 
+## 29.1.0
+
+### Patch Changes
+
+- Updated dependencies [a8192d7]
+- Updated dependencies [a891402]
+  - @hyperlane-xyz/sdk@29.1.0
+  - @hyperlane-xyz/utils@29.1.0
+  - @hyperlane-xyz/core@11.1.0
+
 ## 29.0.1
 
 ### Patch Changes